requirements.txt
# 爬虫相关依赖
selenium==4.8.2
beautifulsoup4==4.12.0
requests==2.28.2
lxml==4.9.3
markdown==3.4.3
html2text==2024.2.26
pyyaml==6.0

# AI分析相关依赖
openai==1.3.5  # 可以根据实际使用的AI服务替换
langchain==0.0.340
python-dotenv==1.0.0

# 工具依赖
webdriver-manager==3.8.5
tqdm==4.66.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0 